+++ packages/sp/src/security/list.ts
@@ -12,9 +12,6 @@ import {
     hasPermissions,
 } from "./funcs";
 
-/**
-* Extend List
-*/
 declare module "../lists/types" {
     interface _List extends ISecurableMethods { }
     interface IList extends ISecurableMethods { }