+++ packages/sp/src/sharepointqueryable.ts
@@ -1,5 +1,5 @@
 import { combine, isUrlAbsolute, assign, jsS, IFetchOptions } from "@pnp/common";
-import { Queryable, IQueryable, invokableFactory, IInvokable } from "@pnp/odata";
+import { Queryable, invokableFactory, IInvokable } from "@pnp/odata";
 import { Logger, LogLevel } from "@pnp/logging";
 import { SPBatch } from "./batch";
 import { metadata } from "./utils/metadata";
@@ -20,7 +20,7 @@ export const spInvokableFactory = <R>(f: any): ISPInvokableFactory<R> => {
  * SharePointQueryable Base Class
  *
  */
-export class _SharePointQueryable<GetType = any> extends Queryable<GetType> implements ISharePointQueryable<GetType> {
+export class _SharePointQueryable<GetType = any> extends Queryable<GetType> {
 
     protected _forceCaching: boolean;
 
@@ -180,21 +180,15 @@ export class _SharePointQueryable<GetType = any> extends Queryable<GetType> impl
         return parent;
     }
 }
-
-export interface ISharePointQueryable<GetType = any> extends IInvokable<GetType>, IQueryable<GetType> {
-    select(...selects: string[]): this;
-    expand(...expands: string[]): this;
-    clone<T extends _SharePointQueryable>(factory: (...args: any[]) => T, additionalPath?: string, includeBatch?: boolean): T;
-    get<T = GetType>(options?: IFetchOptions): Promise<T>;
-}
-export interface _SharePointQueryable extends IInvokable { }
+export interface ISharePointQueryable<GetType = any> extends _SharePointQueryable<GetType>, IInvokable<GetType> { }
+export interface _SharePointQueryable<GetType = any> extends IInvokable<GetType> { }
 export const SharePointQueryable = spInvokableFactory<ISharePointQueryable>(_SharePointQueryable);
 
 /**
  * Represents a REST collection which can be filtered, paged, and selected
  *
  */
-export class _SharePointQueryableCollection<GetType = any[]> extends _SharePointQueryable<GetType> implements ISharePointQueryableCollection<GetType> {
+export class _SharePointQueryableCollection<GetType = any[]> extends _SharePointQueryable<GetType> {
 
     /**
      * Filters the returned collection (https://msdn.microsoft.com/en-us/library/office/fp142385.aspx#bk_supported)
@@ -240,23 +234,15 @@ export class _SharePointQueryableCollection<GetType = any[]> extends _SharePoint
         return this;
     }
 }
-
-export interface ISharePointQueryableCollection<GetType = any[]> extends IInvokable<GetType>, ISharePointQueryable<GetType> {
-    filter(filter: string): this;
-    orderBy(orderBy: string, ascending?: boolean): this;
-    skip(skip: number): this;
-    top(top: number): this;
-    get<T = GetType>(options?: IFetchOptions): Promise<T>;
-}
-
-export interface _SharePointQueryableCollection extends IInvokable { }
+export interface _SharePointQueryableCollection<GetType = any[]> extends IInvokable<GetType> { }
+export interface ISharePointQueryableCollection<GetType = any[]> extends _SharePointQueryableCollection<GetType>, IInvokable<GetType> { }
 export const SharePointQueryableCollection = spInvokableFactory<ISharePointQueryableCollection>(_SharePointQueryableCollection);
 
 /**
  * Represents an instance that can be selected
  *
  */
-export class _SharePointQueryableInstance<GetType = any> extends _SharePointQueryable<GetType> implements ISharePointQueryableInstance<GetType> {
+export class _SharePointQueryableInstance<GetType = any> extends _SharePointQueryable<GetType> {
 
     /**
      * Curries the update function into the common pieces
@@ -273,8 +259,8 @@ export class _SharePointQueryableInstance<GetType = any> extends _SharePointQuer
         }).then((d: Data) => mapper(d, props));
     }
 }
-export interface ISharePointQueryableInstance<GetType = any> extends IInvokable<GetType>, ISharePointQueryable<GetType> { }
-export interface _SharePointQueryableInstance extends IInvokable { }
+export interface ISharePointQueryableInstance<GetType = any> extends _SharePointQueryableInstance<GetType>, IInvokable<GetType> { }
+export interface _SharePointQueryableInstance<GetType = any> extends IInvokable<GetType> { }
 export const SharePointQueryableInstance = spInvokableFactory<ISharePointQueryableInstance>(_SharePointQueryableInstance);
 
 /**