+++ test/common/util.ts
@@ -1,10 +1,11 @@
 import { expect } from "chai";
-import { getCtxCallback } from "..";
-import { dateAdd, combine, getRandomString, getGUID, isFunc, isArray, getAttrValueFromString, assign } from "..";
+import { dateAdd, combine, getRandomString, getGUID, isFunc, isArray, assign, getCtxCallback } from "@pnp/common";
 
-describe("extend", () => {
+// tslint:disable:no-unused-expression
 
-    it("Should extend an object with odd fields", () => {
+describe("extend", function () {
+
+    it("Should extend an object with odd fields", function () {
 
         const o1 = {
             title: "thing",
@@ -19,7 +20,7 @@ describe("extend", () => {
         expect(o).to.deep.eq({ title: "thing", desc: "another" });
     });
 
-    it("Should extend an object with even fields", () => {
+    it("Should extend an object with even fields", function () {
 
         const o1 = {
             desc: "another",
@@ -36,7 +37,7 @@ describe("extend", () => {
         expect(o).to.deep.eq({ desc: "another", title: "thing", bob: "sam", sara: "wendy" });
     });
 
-    it("Should overwrite fields", () => {
+    it("Should overwrite fields", function () {
 
         const o1 = {
             title: "thing",
@@ -51,7 +52,7 @@ describe("extend", () => {
         expect(o).to.deep.eq({ title: "new" });
     });
 
-    it("Should not overwrite fields", () => {
+    it("Should not overwrite fields", function () {
 
         const o1 = {
             title: "thing",
@@ -66,7 +67,7 @@ describe("extend", () => {
         expect(o).to.deep.eq({ title: "thing" });
     });
 
-    it("Should field fields", () => {
+    it("Should field fields", function () {
 
         const o1 = {
             title: "thing",
@@ -83,8 +84,8 @@ describe("extend", () => {
     });
 });
 
-describe("getCtxCallback", () => {
-    it("Should create contextual callback", () => {
+describe("getCtxCallback", function () {
+    it("Should create contextual callback", function () {
 
         class Test {
             constructor(public num = 1) { }
@@ -104,15 +105,15 @@ describe("getCtxCallback", () => {
     });
 });
 
-describe("dateAdd", () => {
-    it("Should add 5 minutes to a date", () => {
+describe("dateAdd", function () {
+    it("Should add 5 minutes to a date", function () {
         const testDate = new Date();
         const checkDate = new Date(testDate.toLocaleString());
         checkDate.setMinutes(testDate.getMinutes() + 5);
         expect(dateAdd(testDate, "minute", 5).getMinutes()).to.eq(checkDate.getMinutes());
     });
 
-    it("Should add 2 years to a date", () => {
+    it("Should add 2 years to a date", function () {
         const testDate = new Date();
         const checkDate = new Date(testDate.toLocaleString());
         checkDate.setFullYear(testDate.getFullYear() + 2);
@@ -120,83 +121,70 @@ describe("dateAdd", () => {
     });
 });
 
-describe("combinePaths", () => {
-    it("Should combine the paths '/path/', 'path2', 'path3' and '/path4' to be path/path2/path3/path4", () => {
+describe("combinePaths", function () {
+    it("Should combine the paths '/path/', 'path2', 'path3' and '/path4' to be path/path2/path3/path4", function () {
         expect(combine("/path/", "path2", "path3", "/path4")).to.eq("path/path2/path3/path4");
     });
 
-    it("Should combine the paths 'http://site/path/' and '/path4/page.aspx' to be http://site/path/path4/page.aspx", () => {
+    it("Should combine the paths 'http://site/path/' and '/path4/page.aspx' to be http://site/path/path4/page.aspx", function () {
         expect(combine("http://site/path/", "/path4/page.aspx")).to.eq("http://site/path/path4/page.aspx");
     });
 
-    it("Should combine the paths null, 'path2', undefined, null and '/path4' to be path2/path4", () => {
+    it("Should combine the paths null, 'path2', undefined, null and '/path4' to be path2/path4", function () {
         expect(combine(null, "path2", undefined, null, "/path4")).to.eq("path2/path4");
     });
 
-    it("Should combine the paths null, 'path2', undefined, \"\", null and '/path4' to be path2/path4", () => {
+    it("Should combine the paths null, 'path2', undefined, \"\", null and '/path4' to be path2/path4", function () {
         expect(combine(null, "path2", undefined, "", null, "/path4")).to.eq("path2/path4");
     });
 
-    it("Should not error with no arguments specified", () => {
+    it("Should not error with no arguments specified", function () {
         expect(combine()).to.eq("");
     });
 });
 
-describe("getRandomString", () => {
-    it("Should produce a random string of length 5", () => {
+describe("getRandomString", function () {
+    it("Should produce a random string of length 5", function () {
         const j = getRandomString(5);
         expect(j).to.be.a("string");
         expect(j).to.have.length(5);
     });
 
-    it("Should produce a random string of length 28", () => {
+    it("Should produce a random string of length 28", function () {
         const j = getRandomString(28);
         expect(j).to.be.a("string");
         expect(j).to.have.length(28);
     });
 });
 
-describe("getGUID", () => {
-    it("Should produce a GUID matching the expected pattern", () => {
+describe("getGUID", function () {
+    it("Should produce a GUID matching the expected pattern", function () {
         expect(getGUID()).to.match(/[a-f0-9]{8}(?:-[a-f0-9]{4}){3}-[a-f0-9]{12}/i);
     });
 });
 
-describe("isFunction", () => {
-    it("Should find that a function is a function", () => {
-        expect(isFunc(() => { return; })).to.be.true;
+describe("isFunction", function () {
+    it("Should find that a function is a function", function () {
+        expect(isFunc(function () { return; })).to.be.true;
     });
 
-    it("Should find that a non-function is not a function", () => {
+    it("Should find that a non-function is not a function", function () {
         expect(isFunc({ val: 0 })).to.be.false;
         expect(isFunc(null)).to.be.false;
         expect(isFunc(undefined)).to.be.false;
     });
 });
 
-describe("isArray", () => {
-    it("Should find that an Array is an Array", () => {
+describe("isArray", function () {
+    it("Should find that an Array is an Array", function () {
         expect(isArray([1, 2, 3, 4])).to.be.true;
     });
 
-    it("Should find that a non-Array is not an Array", () => {
+    it("Should find that a non-Array is not an Array", function () {
         expect(isArray(null)).to.be.false;
         expect(isArray("")).to.be.false;
         expect(isArray(3)).to.be.false;
         expect(isArray({})).to.be.false;
         expect(isArray(undefined)).to.be.false;
-    });
-});
-
-describe("getAttrValueFromString", () => {
-
-    it("Should correctly parse attribute values", () => {
-        expect(getAttrValueFromString(`<thing att='value' />`, "att")).to.eq("value");
-        expect(getAttrValueFromString(`<thing att="value1293.?,/\|!@#$%^&*()[]{}" />`, "att")).to.eq("value1293\\.\\?,/\\|!@#\\$%\\^&\\*\\(\\)\\[\\]\\{\\}");
-        expect(getAttrValueFromString(`<thing att  =  'value' />`, "att")).to.eq("value");
-        expect(getAttrValueFromString(`<thing att='value"' />`, "att")).to.eq("value\"");
-        expect(getAttrValueFromString(`<thing att="value'" />`, "att")).to.eq("value'");
-        expect(getAttrValueFromString(`<thing att='value value' att2="something" />`, "att")).to.eq("value value");
-        expect(getAttrValueFromString(`<thing att='value'></thing>`, "att")).to.eq("value");
     });
 });